Implement track randomizer

+++ core/player/common/src/main/java/net/sigmabeta/chipbox/player/common/GeneratorEvent.kt
@@ -8,7 +8,7 @@ sealed class GeneratorEvent {
     data class Audio(
         val bufferNumber: Int,
         val timestampMillis: Double,
-        val timestampFrame: Long,
+        val timestampFrame: Int,
         val sampleRate: Int,
         val data: ShortArray
     ) : GeneratorEvent()
